fix: Revert PR 5110
PR #5110 breaks version streams BDD tests, where they use a specific
ref from the version stream PR.

```
Deleting and cloning the Jenkins X versions repo
Cloning the Jenkins X versions repo https://github.com/jenkins-x/jenkins-x-versions.git with revision 386ca964ed73fa71dd8f1f6c0acd65ea62b20070 to /builder/home/.**/jenkins-x-versions
error: there was a problem creating a version resolver from versions stream repository https://github.com/jenkins-x/jenkins-x-versions.git and ref 386ca964ed73fa71dd8f1f6c0acd65ea62b20070: running git describe 386ca964ed73fa71dd8f1f6c0acd65ea62b20070 --abbrev=0: running git describe 386ca964ed73fa71dd8f1f6c0acd65ea62b20070 --abbrev=0 --contains: failed to run 'git describe 386ca964ed73fa71dd8f1f6c0acd65ea62b20070 --abbrev=0 --contains' command in directory '/builder/home/.**/jenkins-x-versions', output: 'fatal: cannot describe '386ca964ed73fa71dd8f1f6c0acd65ea62b20070''
ERROR: Error: Command failed  ** boot -b
error: exit status 1
```
