+++ pkg/cmd/opts/team_settings.go
@@ -36,10 +36,7 @@ const (
 // TeamSettings returns the team settings
 func (o *CommonOptions) TeamSettings() (*v1.TeamSettings, error) {
 	_, teamSettings, err := o.DevEnvAndTeamSettings()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-	return teamSettings, nil
+	return teamSettings, err
 }
 
 // DevEnvAndTeamSettings returns the Dev Environment and Team settings
