+++ pkg/cmd/opts/upgrade/upgrade_ingress.go
@@ -18,7 +18,7 @@ import (
 	"github.com/jenkins-x/jx/pkg/kube"
 	"github.com/jenkins-x/jx/pkg/log"
 	"github.com/jenkins-x/jx/pkg/util"
-	survey "gopkg.in/AlecAivazis/survey.v1"
+	"gopkg.in/AlecAivazis/survey.v1"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/api/extensions/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -72,8 +72,7 @@ func (o *UpgradeIngressOptions) Run() error {
 	// wizard to ask for config values
 	err = o.confirmExposecontrollerConfig()
 	if err != nil {
-		return errors.Wrap(err, ""+
-			"configure exposecontroller")
+		return errors.Wrap(err, "configure exposecontroller")
 	}
 
 	// confirm values
