+++ pkg/cmd/step/verify/step_verify_environments_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	v1 "github.com/jenkins-x/jx/pkg/apis/jenkins.io/v1"
-	"github.com/jenkins-x/jx/pkg/boot"
 	"github.com/jenkins-x/jx/pkg/cmd/clients/fake"
 	"github.com/jenkins-x/jx/pkg/cmd/opts"
 	"github.com/jenkins-x/jx/pkg/cmd/opts/step"
@@ -137,15 +136,15 @@ func Test_IsJXBoot(t *testing.T) {
 }
 
 func Test_ReadEnvironment(t *testing.T) {
-	origConfigRepoURL, foundConfigRepoURLEnvKey := os.LookupEnv(boot.ConfigRepoURLEnvVarName)
-	origConfigRepoRef, foundConfigRepoRefEnvKey := os.LookupEnv(boot.ConfigBaseRefEnvVarName)
+	origConfigRepoURL, foundConfigRepoURLEnvKey := os.LookupEnv(configRepoURLEnvKey)
+	origConfigRepoRef, foundConfigRepoRefEnvKey := os.LookupEnv(configRepoRefEnvKey)
 	defer func() {
 		if foundConfigRepoURLEnvKey {
-			_ = os.Setenv(boot.ConfigRepoURLEnvVarName, origConfigRepoURL)
+			_ = os.Setenv(configRepoURLEnvKey, origConfigRepoURL)
 		}
 
 		if foundConfigRepoRefEnvKey {
-			_ = os.Setenv(boot.ConfigBaseRefEnvVarName, origConfigRepoRef)
+			_ = os.Setenv(configRepoRefEnvKey, origConfigRepoRef)
 		}
 	}()
 
@@ -158,26 +157,26 @@ func Test_ReadEnvironment(t *testing.T) {
 		errorString string
 	}{
 		{"https://github.com/jenkins-x/jenkins-x-boot-config", "master", false, ""},
-		{"https://github.com/jenkins-x/jenkins-x-boot-config", "", true, "the environment variable CONFIG_BASE_REF must be specified"},
-		{"", "master", true, "the environment variable CONFIG_REPO_URL must be specified"},
-		{"", "", true, "[the environment variable CONFIG_REPO_URL must be specified, the environment variable CONFIG_BASE_REF must be specified]"},
+		{"https://github.com/jenkins-x/jenkins-x-boot-config", "", true, "the environment variable BASE_CONFIG_REF must be specified"},
+		{"", "master", true, "the environment variable REPO_URL must be specified"},
+		{"", "", true, "[the environment variable REPO_URL must be specified, the environment variable BASE_CONFIG_REF must be specified]"},
 	}
 
 	for _, testCase := range tests {
 		t.Run(fmt.Sprintf("%s-%s", testCase.url, testCase.ref), func(t *testing.T) {
 			if testCase.url == "" {
-				err := os.Unsetenv(boot.ConfigRepoURLEnvVarName)
+				err := os.Unsetenv(configRepoURLEnvKey)
 				assert.NoError(t, err)
 			} else {
-				err := os.Setenv(boot.ConfigRepoURLEnvVarName, testCase.url)
+				err := os.Setenv(configRepoURLEnvKey, testCase.url)
 				assert.NoError(t, err)
 			}
 
 			if testCase.ref == "" {
-				err := os.Unsetenv(boot.ConfigBaseRefEnvVarName)
+				err := os.Unsetenv(configRepoRefEnvKey)
 				assert.NoError(t, err)
 			} else {
-				err := os.Setenv(boot.ConfigBaseRefEnvVarName, testCase.ref)
+				err := os.Setenv(configRepoRefEnvKey, testCase.ref)
 				assert.NoError(t, err)
 
 			}
