+++ pkg/gits/git_cli_test.go
@@ -756,77 +756,3 @@ func TestGitCLI_Remotes(t *testing.T) {
 		})
 	}
 }
-
-func TestDescribe(t *testing.T) {
-	tests := []struct {
-		initFn   func(dir string, gitter gits.Gitter) error
-		name     string
-		want     []string
-		wantErr  bool
-		contains bool
-		abbrev   string // Number of hex digits from object name, defaults to 7
-	}{
-		{
-			initFn: func(dir string, git gits.Gitter) error {
-				err := git.Init(dir)
-				assert.NoError(t, err)
-
-				err = git.AddCommit(dir, "Initial Commit")
-				assert.NoError(t, err)
-				err = git.CreateTag(dir, "v0.0.1", "First Tag")
-				assert.NoError(t, err)
-				return err
-			},
-			name:     "Valid commit and tag, !contains, abbrev=default",
-			wantErr:  false,
-			contains: false,
-			abbrev:   "",
-		},
-		{
-			initFn: func(dir string, git gits.Gitter) error {
-				var err error
-				assert.NoError(t, err)
-				err = git.Init(dir)
-				assert.NoError(t, err)
-
-				err = git.AddCommit(dir, "Initial Commit")
-				assert.NoError(t, err)
-				return err
-			},
-			name:     "Commit but no tag, !contains, abbrev=default",
-			wantErr:  true,
-			contains: false,
-			abbrev:   "",
-		},
-	}
-
-	for _, test := range tests {
-		dir, err := ioutil.TempDir("", "")
-		assert.NoError(t, err)
-		defer func() {
-			os.RemoveAll(dir)
-		}()
-
-		gitCli := gits.NewGitCLI()
-
-		err = test.initFn(dir, gitCli)
-		assert.NoError(t, err)
-		parts1, parts2, err := gitCli.Describe(
-			dir,
-			test.contains,
-			"HEAD",
-			test.abbrev,
-		)
-
-		if !test.wantErr {
-			assert.NoError(t, err)
-			assert.NotNil(t, parts1)
-			assert.Equal(t, "", parts2)
-		} else {
-			assert.Error(t, err)
-			assert.Equal(t, "", parts1)
-			assert.Equal(t, "", parts2)
-		}
-	}
-
-}
