+++ pkg/gits/git_fake.go
@@ -655,8 +655,3 @@ func (g *GitFake) CherryPick(dir string, commit string) error {
 func (g *GitFake) CherryPickTheirs(dir string, commit string) error {
 	return nil
 }
-
-// Describe does a git describe of commitish, optionally adding the abbrev arg if not empty
-func (g *GitFake) Describe(dir string, contains bool, commitish string, abbrev string) (string, string, error) {
-	return "", "", nil
-}
