+++ pkg/gits/git_local.go
@@ -508,8 +508,3 @@ func (g *GitLocal) CherryPick(dir string, commit string) error {
 func (g *GitLocal) CherryPickTheirs(dir string, commit string) error {
 	return g.GitCLI.CherryPickTheirs(dir, commit)
 }
-
-// Describe does a git describe of commitish, optionally adding the abbrev arg if not empty
-func (g *GitLocal) Describe(dir string, contains bool, commitish string, abbrev string) (string, string, error) {
-	return g.GitCLI.Describe(dir, false, commitish, abbrev)
-}
